# public/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hub Sí Viajo - API de Paquetes",
    "description": "API para gestionar paquetes turísticos del Hub Sí Viajo. Permite listar, filtrar y actualizar paquetes de forma programática. Con `include=full` se obtiene el itinerario completo: destinos, transporte con segmentos, hoteles con habitaciones e imágenes, transfers, tours, autos, tickets, seguros, imágenes y precios.",
    "version": "2.0.0",
    "contact": {
      "name": "Soporte Hub Sí Viajo",
      "url": "https://hub.siviajo.com"
    }
  },
  "servers": [
    {
      "url": "https://hub.siviajo.com/api",
      "description": "Producción"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
    "/packages": {
      "get": {
        "summary": "Listar paquetes",
        "description": "Obtiene una lista de paquetes con filtros opcionales y paginación. Usar `include=full` para obtener el itinerario completo con todas las relaciones.",
        "operationId": "getPackages",
        "tags": ["Paquetes"],
        "parameters": [
          {
            "name": "tc_package_id",
            "in": "query",
            "description": "Filtrar por ID de TravelCompositor",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 12345
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filtrar por estado del paquete",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["imported", "reviewing", "approved", "in_design", "in_marketing", "published", "expired"],
              "example": "in_marketing"
            }
          },
          {
            "name": "needs_quote",
            "in": "query",
            "description": "Filtrar por necesidad de cotización manual",
            "required": false,
            "schema": {
              "type": "boolean",
              "example": true
            }
          },
          {
            "name": "active",
            "in": "query",
            "description": "Filtrar por estado activo en TravelCompositor",
            "required": false,
            "schema": {
              "type": "boolean",
              "example": true
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Buscar en el título del paquete",
            "required": false,
            "schema": {
              "type": "string",
              "example": "cancun"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Cantidad de resultados por página",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 500,
              "example": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Desplazamiento para paginación",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0,
              "example": 0
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Campo para ordenar resultados",
            "required": false,
            "schema": {
              "type": "string",
              "default": "created_at",
              "enum": ["created_at", "departure_date", "current_price_per_pax", "title", "last_sync_at"],
              "example": "departure_date"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Dirección del ordenamiento",
            "required": false,
            "schema": {
              "type": "string",
              "default": "desc",
              "enum": ["asc", "desc"],
              "example": "asc"
            }
          },
          {
            "name": "include",
            "in": "query",
            "description": "Nivel de detalle. Usar 'full' para incluir itinerario completo: destinos, transportes con segmentos, hoteles con habitaciones e imágenes, transfers, tours, autos, tickets, seguros, imágenes del paquete, precios por servicio, desglose de costos e historial de precios",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["full"],
              "example": "full"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de paquetes obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackagesListResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado - API Key inválida o faltante",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Actualizar paquetes en lote",
        "description": "Actualiza múltiples paquetes con los mismos valores.",
        "operationId": "updatePackages",
        "tags": ["Paquetes"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Paquetes actualizados exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkUpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida - IDs o updates faltantes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key proporcionada por el administrador del Hub"
      }
    },
    "schemas": {
      "Package": {
        "type": "object",
        "description": "Paquete turístico básico (sin include=full)",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID único en la base de datos del Hub",
            "example": 1
          },
          "tc_package_id": {
            "type": "integer",
            "description": "ID del paquete en TravelCompositor",
            "example": 12345
          },
          "title": {
            "type": "string",
            "description": "Título del paquete",
            "example": "Cancún All Inclusive 7 noches"
          },
          "large_title": {
            "type": "string",
            "nullable": true,
            "description": "Título largo/descriptivo",
            "example": "Cancún All Inclusive 7 noches con vuelo directo"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Descripción del paquete"
          },
          "image_url": {
            "type": "string",
            "nullable": true,
            "description": "URL de la imagen principal",
            "example": "https://cdn.siviajo.com/images/package-123.jpg"
          },
          "external_reference": {
            "type": "string",
            "nullable": true,
            "description": "Referencia externa"
          },
          "tc_creation_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Fecha de creación en TravelCompositor"
          },
          "departure_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "Fecha de salida",
            "example": "2025-03-15"
          },
          "date_range_start": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "Inicio del rango de fechas disponibles",
            "example": "2025-03-01"
          },
          "date_range_end": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "Fin del rango de fechas disponibles",
            "example": "2025-03-31"
          },
          "original_price_per_pax": {
            "type": "number",
            "nullable": true,
            "description": "Precio original por persona"
          },
          "current_price_per_pax": {
            "type": "number",
            "description": "Precio actual por persona",
            "example": 1500.00
          },
          "total_price": {
            "type": "number",
            "nullable": true,
            "description": "Precio total del paquete"
          },
          "currency": {
            "type": "string",
            "description": "Moneda del precio",
            "example": "USD"
          },
          "price_variance_pct": {
            "type": "number",
            "description": "Porcentaje de variación del precio"
          },
          "adults_count": {
            "type": "integer",
            "description": "Cantidad de adultos",
            "example": 2
          },
          "children_count": {
            "type": "integer",
            "description": "Cantidad de niños",
            "example": 0
          },
          "infants_count": {
            "type": "integer",
            "description": "Cantidad de infantes",
            "example": 0
          },
          "nights_count": {
            "type": "integer",
            "description": "Cantidad de noches",
            "example": 7
          },
          "destinations_count": {
            "type": "integer",
            "description": "Cantidad de destinos",
            "example": 1
          },
          "transports_count": {
            "type": "integer",
            "description": "Cantidad de transportes"
          },
          "hotels_count": {
            "type": "integer",
            "description": "Cantidad de hoteles"
          },
          "transfers_count": {
            "type": "integer",
            "description": "Cantidad de transfers"
          },
          "cars_count": {
            "type": "integer",
            "description": "Cantidad de autos"
          },
          "tickets_count": {
            "type": "integer",
            "description": "Cantidad de tickets/entradas"
          },
          "tours_count": {
            "type": "integer",
            "description": "Cantidad de tours"
          },
          "tc_active": {
            "type": "boolean",
            "description": "Si está activo en TravelCompositor",
            "example": true
          },
          "status": {
            "type": "string",
            "enum": ["imported", "reviewing", "approved", "in_design", "in_marketing", "published", "expired"],
            "description": "Estado del paquete en el flujo de trabajo",
            "example": "in_marketing"
          },
          "needs_manual_quote": {
            "type": "boolean",
            "description": "Si requiere cotización manual",
            "example": false
          },
          "send_to_design": {
            "type": "boolean",
            "description": "Marcado para diseño",
            "example": true
          },
          "send_to_design_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Fecha de envío a diseño"
          },
          "design_completed": {
            "type": "boolean",
            "description": "Diseño completado",
            "example": true
          },
          "design_completed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Fecha de completado de diseño"
          },
          "send_to_marketing": {
            "type": "boolean",
            "description": "Marcado para marketing",
            "example": true
          },
          "send_to_marketing_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Fecha de envío a marketing"
          },
          "marketing_completed": {
            "type": "boolean",
            "description": "Marketing completado",
            "example": false
          },
          "marketing_completed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Fecha de completado de marketing"
          },
          "seo_title": {
            "type": "string",
            "nullable": true,
            "description": "Título SEO"
          },
          "seo_description": {
            "type": "string",
            "nullable": true,
            "description": "Descripción SEO"
          },
          "ai_description": {
            "type": "string",
            "nullable": true,
            "description": "Descripción generada por IA"
          },
          "in_sitemap": {
            "type": "boolean",
            "description": "Incluido en el sitemap"
          },
          "origin_code": {
            "type": "string",
            "nullable": true,
            "description": "Código del origen"
          },
          "origin_name": {
            "type": "string",
            "nullable": true,
            "description": "Nombre del origen"
          },
          "origin_country": {
            "type": "string",
            "nullable": true,
            "description": "País del origen"
          },
          "themes": {
            "type": "array",
            "nullable": true,
            "items": { "type": "string" },
            "description": "Temáticas del paquete"
          },
          "tc_idea_url": {
            "type": "string",
            "nullable": true,
            "description": "URL de la idea en TravelCompositor"
          },
          "last_requote_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Última recotización"
          },
          "requote_status": {
            "type": "string",
            "nullable": true,
            "description": "Estado de recotización"
          },
          "available_dates": {
            "type": "array",
            "nullable": true,
            "items": { "type": "string" },
            "description": "Fechas disponibles"
          },
          "package_type": {
            "type": "string",
            "nullable": true,
            "description": "Tipo de paquete"
          },
          "min_nights": {
            "type": "integer",
            "nullable": true,
            "description": "Noches mínimas"
          },
          "max_nights": {
            "type": "integer",
            "nullable": true,
            "description": "Noches máximas"
          },
          "is_refundable": {
            "type": "boolean",
            "description": "Si es reembolsable"
          },
          "tc_username": {
            "type": "string",
            "nullable": true,
            "description": "Usuario de TravelCompositor"
          },
          "cancellation_policy": {
            "type": "string",
            "nullable": true,
            "description": "Política de cancelación"
          },
          "ads_created_count": {
            "type": "integer",
            "description": "Cantidad de anuncios creados",
            "example": 3
          },
          "imported_at": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de importación"
          },
          "last_sync_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Última sincronización con TC",
            "example": "2025-01-20T08:00:00Z"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de creación en el Hub",
            "example": "2025-01-15T10:30:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de última actualización"
          }
        }
      },
      "PackageDestination": {
        "type": "object",
        "description": "Destino del paquete",
        "properties": {
          "id": { "type": "integer" },
          "package_id": { "type": "integer" },
          "destination_code": { "type": "string", "example": "CUN" },
          "destination_name": { "type": "string", "example": "Cancún" },
          "country": { "type": "string", "nullable": true, "example": "México" },
          "country_code": { "type": "string", "nullable": true, "example": "MX" },
          "from_day": { "type": "integer", "nullable": true, "description": "Día de inicio en el destino" },
          "to_day": { "type": "integer", "nullable": true, "description": "Día de fin en el destino" },
          "nights": { "type": "integer", "nullable": true, "description": "Noches en el destino" },
          "latitude": { "type": "number", "nullable": true },
          "longitude": { "type": "number", "nullable": true },
          "recommended_airport_code": { "type": "string", "nullable": true },
          "recommended_airport_name": { "type": "string", "nullable": true },
          "description": { "type": "string", "nullable": true },
          "sort_order": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "PackageTransfer": {
        "type": "object",
        "description": "Transfer del paquete",
        "properties": {
          "id": { "type": "integer" },
          "package_id": { "type": "integer" },
          "tc_transfer_id": { "type": "string", "nullable": true },
          "tc_provider_code": { "type": "string", "nullable": true },
          "supplier_name": { "type": "string", "nullable": true },
          "day": { "type": "integer", "nullable": true },
          "transfer_type": { "type": "string", "nullable": true, "example": "PRIVATE" },
          "from_name": { "type": "string", "nullable": true, "example": "Aeropuerto Cancún" },
          "from_latitude": { "type": "number", "nullable": true },
          "from_longitude": { "type": "number", "nullable": true },
          "to_name": { "type": "string", "nullable": true, "example": "Grand Oasis Cancun" },
          "to_latitude": { "type": "number", "nullable": true },
          "to_longitude": { "type": "number", "nullable": true },
          "vehicle_type": { "type": "string", "nullable": true },
          "service_type": { "type": "string", "nullable": true },
          "product_type": { "type": "string", "nullable": true },
          "datetime": { "type": "string", "format": "date-time", "nullable": true },
          "duration_minutes": { "type": "integer", "nullable": true },
          "description": { "type": "string", "nullable": true },
          "image_url": { "type": "string", "nullable": true },
          "net_price": { "type": "number", "nullable": true },
          "total_price": { "type": "number", "nullable": true },
          "currency": { "type": "string" },
          "mandatory": { "type": "boolean" },
          "adults_count": { "type": "integer" },
          "children_count": { "type": "integer" },
          "infants_count": { "type": "integer" },
          "sort_order": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "PackagePriceHistory": {
        "type": "object",
        "description": "Historial de precios del paquete",
        "properties": {
          "id": { "type": "integer" },
          "package_id": { "type": "integer" },
          "price_per_pax": { "type": "number" },
          "total_price": { "type": "number", "nullable": true },
          "currency": { "type": "string" },
          "previous_price": { "type": "number", "nullable": true },
          "variance_amount": { "type": "number", "nullable": true },
          "variance_pct": { "type": "number", "nullable": true },
          "recorded_at": { "type": "string", "format": "date-time" }
        }
      },
      "PackageWithRelations": {
        "allOf": [
          { "$ref": "#/components/schemas/Package" },
          {
            "type": "object",
            "description": "Paquete completo con itinerario (con include=full)",
            "properties": {
              "package_destinations": {
                "type": "array",
                "description": "Destinos del paquete",
                "items": { "$ref": "#/components/schemas/PackageDestination" }
              },
              "package_transports": {
                "type": "array",
                "description": "Transportes del paquete con segmentos de vuelo",
                "items": { "$ref": "#/components/schemas/PackageTransport" }
              },
              "package_hotels": {
                "type": "array",
                "description": "Hoteles con habitaciones e imágenes",
                "items": { "$ref": "#/components/schemas/PackageHotel" }
              },
              "package_transfers": {
                "type": "array",
                "description": "Transfers del paquete",
                "items": { "$ref": "#/components/schemas/PackageTransfer" }
              },
              "package_closed_tours": {
                "type": "array",
                "description": "Tours cerrados",
                "items": { "$ref": "#/components/schemas/PackageClosedTour" }
              },
              "package_cars": {
                "type": "array",
                "description": "Alquiler de autos",
                "items": { "$ref": "#/components/schemas/PackageCar" }
              },
              "package_tickets": {
                "type": "array",
                "description": "Tickets y entradas",
                "items": { "$ref": "#/components/schemas/PackageTicket" }
              },
              "package_images": {
                "type": "array",
                "description": "Imágenes del paquete",
                "items": { "$ref": "#/components/schemas/PackageImage" }
              },
              "package_insurances": {
                "type": "array",
                "description": "Seguros del paquete",
                "items": { "$ref": "#/components/schemas/PackageInsurance" }
              },
              "package_service_prices": {
                "type": "array",
                "description": "Desglose de precios por servicio",
                "items": { "$ref": "#/components/schemas/PackageServicePrice" }
              },
              "package_cost_breakdown": {
                "type": "array",
                "description": "Desglose de costos totales",
                "items": { "$ref": "#/components/schemas/PackageCostBreakdown" }
              },
              "package_price_history": {
                "type": "array",
                "description": "Historial de cambios de precio",
                "items": { "$ref": "#/components/schemas/PackagePriceHistory" }
              }
            }
          }
        ]
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "Total de registros",
            "example": 150
          },
          "limit": {
            "type": "integer",
            "description": "Límite por página",
            "example": 50
          },
          "offset": {
            "type": "integer",
            "description": "Desplazamiento actual",
            "example": 0
          },
          "hasMore": {
            "type": "boolean",
            "description": "Si hay más páginas disponibles",
            "example": true
          }
        }
      },
      "PackagesListResponse": {
        "type": "object",
        "properties": {
          "packages": {
            "type": "array",
            "description": "Sin include=full devuelve Package básico, con include=full devuelve PackageWithRelations",
            "items": {
              "$ref": "#/components/schemas/PackageWithRelations"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
      "BulkUpdateRequest": {
        "type": "object",
        "required": ["ids", "updates"],
        "properties": {
          "ids": {
            "type": "array",
            "description": "Array de IDs de paquetes a actualizar",
            "items": {
              "type": "integer"
            },
            "example": [1, 2, 3]
          },
          "updates": {
            "type": "object",
            "description": "Campos a actualizar",
            "properties": {
              "status": {
                "type": "string",
                "enum": ["imported", "reviewing", "approved", "in_design", "in_marketing", "published", "expired"]
              },
              "send_to_design": {
                "type": "boolean"
              },
              "design_completed": {
                "type": "boolean"
              },
              "send_to_marketing": {
                "type": "boolean"
              },
              "marketing_completed": {
                "type": "boolean"
              },
              "needs_manual_quote": {
                "type": "boolean"
              },
              "seo_title": {
                "type": "string"
              },
              "seo_description": {
                "type": "string"
              },
              "ai_description": {
                "type": "string"
              },
              "in_sitemap": {
                "type": "boolean"
              }
            }
          }
        }
      },
      "BulkUpdateResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "updated": {
            "type": "integer",
            "description": "Cantidad de registros actualizados",
            "example": 3
          },
          "packages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Package"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Mensaje de error",
            "example": "No autorizado"
          }
        }
      }
    }
  },
...
}